refactor: Move files queries
